Defer Event Registration
OnStart would normally be fine, but RFMEC / B9PS end up firing OnVesselModified very often during OnStart as well.  This leads to a lot of unnecessary cache recomputation during startup.  Wait until the first time the data is needed in Update to register and initialize.
Skip redundant check for IsEngine() and !IsSeparatron() if we are iterating an engines cache, those are already validated.