Fix PVG with upper stage solids (Castor 30XL)
- Adds a new MaxThrust field to the stage stats display

This is computed as the MaxThrust of all the engines burning
at the end of the stage.  This is deliberately to avoid counting
ullage solids that burnout after a few seconds.  It may, however,
be viewed as being slightly misleading.  But the purpose of this
field is really for PVG to have a single number for the thrust of the
stage which is more realistic than the current behavior where it
gets the initial thrust of the engine including the thrust curve.

- Converts PVG to use MaxThrust + ISP + MassDelta to get total
burntime instead of the other way around.  PVGs computed burntime
for a Castor 30 XL is now 2mins instead of 8 mins.

- The displayed burntime, TWRs and SLTs are still wildly off, those
need to be fixed more deeply in the code which analyzes the
engine and produces the mass flow rate at conditions.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>