Tweak tranfer planner SQP settings
Adjust the diffstep way down, experimentation with Matlab indicates this
may not hinder convergence as much as the alglib author suggests.

Adjust EPSX upwards so stopping conditions are more lenient.

This worked on my one test so far with a high level of accuracy on a
target from a horrible parking orbit on Earth to Saturn with a periapsis
altitude of 2010km (2,010,006m actual).

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>