Fix TimeToPlane function
Calculates what the LAN would be if launching right now using Napier's
rules; then finds the time until this LAN would be equal to the LAN
target. Correctly works for all inclinations and latitudes, with the
convention that a negative inclination specifies the southern launch
window.

Verified in GeoGebra (https://www.geogebra.org/calculator/uxzbpdpt);
also ran in-game tests in RSS from KSC (latitude 28°) and Woomera
(latitude -30°) for inclinations:
 - -180° < inc < -90° (retrograde orbit, southern window)
 - -90° < inc < -|latitude| (prograde orbit, southern window)
 - |latitude| < inc < 90° (prograde orbit, northern window)
 - 90° < inc < 180° (retrograde orbit, northern window)
 - |inc| < |latitude| (prograde orbit, with a dog-leg)
 - |180° - inc| < |latitude| (retrograde orbit, with a dog-leg)
 - inc = 90° (polar orbit, northern window)
 - inc = -90° (polar orbit, southern window)

The test for inclination = -90° _fails_ for both KSC and Woomera - PVG
fails with "Unreachable even with infinite Isp". I suspect that this is
caused by a singularity in the boundary conditions, but have not been
able to verify this.

Also fixes a bug where for clockwise-rotating planets (with
rotationPeriod < 0), the returned time was negative. For example, Venus
in RSS has this property.