Fix PVG launching into the wrong plane
When both a LAN and inclination are specified and the inclination is
negative, PVG launches into the wrong plane.

Inclination is always positive in the physical sense; however, MechJeb's
ascent guidance uses a negative inclination to signify that the the
southern launch window should be used (with a positive inclination using
the northern launch window).

In the boundary conditions with a fixed LAN, also the sine of the
inclination is used to determine the orbital plane. A negative
inclination would result in an orbital plane that has it's DN at the
requested AN and vice-versa.

In the boundary conditions with a free LAN, only the cosine of the
inclination is used; the result is that the choice of north vs south
depends on which local minimum the optimizer finds; sign of the
inclination has no influence. Changing this such that the optimizer
always lands at the correct local minimum would require changes to the
boundary conditions.