Use sign of inclination to select window for PVG
If the inclination is negative, use the southern launch window; if it is
positive, use the northern launch window. Previously, the nearest of the
two launch windows was taken.

Cannot be done for "Launch into plane of target", since the inclination
for that mode is taken directly from the target orbit and is thus always
positive.