Fix PVG launch into plane of target
There were actually two related issues:

1. MJ uses a negative inclination sign as a flag that the southern
   launch window should be chosen; however, this flag is not set by the
   `MinimumTimeToPlane` function, thereby making the heading used for
   the initial guess fed into the optimizer wrong.

2. The boundary conditions in PVG assume that the inclination is always
   positive, which caused PVG to try and launch into the wrong plane
   (LAN offset by 180°, or AN and DN swapped).

The fix is to change the `MinimumTimeToPlane` function to feed-back
whether the chosen launch window is the north or south one, and to
update the sign of the inclination accordingly. This is then fed into
the HeadingForInclination function to determine the heading for the
initial guess. After this, the inclination is stripped of its sign flag
before being fed into the PVG boundary condition functions.

The reason that it worked previously most of the time, is that for lower
inclinations, the heading error between north and south is small enough
that the optimizer can correct it. For near-polar inclinations, the
error grows to 180° and the optimizer failed to find a solution.

With these fixes, "Launch into plane of target", "Launch into target
LAN" and "Launch to manual LAN" should all work correctly for all
cominations of inclination and LAN.