Respect the user-provided inclination sign
Instead of warping to the nearest of the northern or southern windows,
warp to the northern window if inclination is positive and southern if
inclination is negative.

Only in the case of "Launch into LAN of target" and "Launch into manual
LAN"; the case "Launch into plane of target" keeps the previous
behavior: for this case, the inclination input by the user is ignored in
favor of the inclination of the target (which is always positive), so it
seems weird to ignore the user input _except_ for the sign.