Fix FuelFlowSim for NO_FLOW resources in Engines
This fixes a bug affecting Delta IV Heavy setup with asparagus staging.  The RS68 engines have had Ablator added to them as an additional "fuel" which is NO_FLOW.

The old check CanDrawResource() which was used to determine if staging was allowed was not aware of fuel modes and that caused it to check if any part in the crossfeedset had the fuel even for NO_FLOW fuels.  This meant that after the side boosters flamed out that the center core would still be burning Ablator and it would that Ablator "accessible" in its crossfeed set coming from the side engines (because of asparagus).

This may also affect SRBs since those fuels are NO_FLOW as well?

However, I think it requires something like Asparagus staging And given that RO doesn't allow crossfeed through decouplers or fuel pipes it may be a pretty narrow bugfix.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>