Fix StagingController for NO_FLOW fuels
Companion to #1612 to fix autostaging.

Had to sort of invert the algorithm to loop over all the propellants in
all the active engines so that I could grab the Mode off of the engine
propellant (seems not to be associated with the resources on the tank).

This worked on my test Delta IV Heavy on the first try...  Makes me
wonder what I broke...

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>