Hang state off the IVP object
This reduces the method signatures but is less functional.

I was starting down the road of passing t, tnew, y, ynew and
dy, dynew to half of the callback methods, which starts to
look like litter.

Amazingly this fixes the Shepperd tests which seems to have been
a real bug fixed by the new Habs/_habsNext handling (probably in
better handling of Tnew and the isWithin() comparison when
pulling off the interpolated values?)

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>