Update MechJebLib copyright headers
This utilizes the SPDX-License-Identifier to be more permissive and
legally clear and extend the license under MechJebLib.

It follows https://cr.yp.to/publicdomain.html in order to place the
work under the public domain.

The other licenses are also included to satisfy license checkers and
to provide a deliberate fallback for any concerns over the public
domain declaration.  Worst case this should be equivalent to dual
licensing under 0BSD and LGPL2.1+.

The License Identifier also clears up my poor wording using "and" under
the dual license when the intent was to make it at the choice of the
user, not to have both always applied (which doesn't make any sense to
me which is why I didn't think it needed any clarification)

This work is hereby placed into the public domain.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>