PVG: Use Analytic for bootstrap and Integrator for converged
Makes bootstrapping fast and then makes it accurate.

Also can find issues with the integrator by having it be the
actual second pass through.  It should be very definitive bug
reports as long as people describe it properly or show a video.

Had to initialize the Vn's to zeros because otherwise weird things
happen due to some kind of reuse bug.  Someting is getting +='d or
something like that without being zero'd.  Really that's hard
enough to track down that zero'ing everything makes a lot of sense.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>