Fix autostaging for asparagus boosters
In this loop, "continue" means "this condition doesn't prevent staging
(though another might)"; "return true" means "this tank still contains
resources; don't stage it away yet".

Given that interpretation, the condition on this line is reversed: when
there's no more of a resource that can be pulled from this tank, then it
is not a problem if we stage the tank away (irrespective of whether
there's any active engine that can pull it - which is what the rest of
the function checks).

In PR #1613 the other checks in the loop were inverted, but this one was missed.

Fixes #1658.