Fix Fairing Separation logic
Fairing stages now must be:

- if a PF fairing is in the stage, all the parts in the stage must
  be PF payload fairings.

- if a stage has only decopulers, without children, that aren't launch
  clamps then it is considered a fairing stage.

this fixes the bug where any PF fairing in a stage would turn it into
a fairing stage.

this also fixes things like RSB fairing decouplers, which are just
normal decouplers and it is hard to identify them uniquely.

this may identify non-PF interstage fairings that are in a stage by
themselves as payload fairings.  the workaround to that is to add the
stack decoupler to that stage and stage them together (or an engine,
or any other part which will prevent them being identified as a
payload fairing).

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>