Fix both checks that use IndexForKSPStage
Changes the API to more sensibly just use the coast/burn boolean as
a matcher, so we try to find exactly what we're looking for.

Also make sure that we're continuously updating the _blockOptimizerUntilTime
code so that we get the 5 seconds of suspending the optimizer correctly
between the coast and the burn when we're doing coast-before and
therefore don't hit the staging callback.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>