Landing autopilot style cleanup
This may be breaking due to changing methods to private and changing
the name of public methods.

If anything has been made private that is actually getting used just
drop in a bug report to revert it and it can be made public with a
[UsedImplicitly] added to make IDEs happy.  It isn't intended to
force people to beat on it with reflection calls to private APIs.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>