Possibly breaking cleanup of MechJebCore
The renaming here is intended to be breaking, mods will need to
update.  I tried to keep public APIs that looked deliberately
public to be [UsedImplicitly] and visible, if the visibility
changed of something that should be made public that is a bug
and it can be reverted back to public.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>