Update to C# 9.0, add partial suicide burn overhaul
This is kinda weird, its got a half-finished fully numerical suicide
burn overhaul I did which needs to be finished, also bumps the version
to C# 9 and fixes a problem with the FlightRecorder using the 'record'
newly reserved word, along with some support for BackgroundJob in the
MechJebUtils class.

What I really want is that support for BackgroundJobs to use with the
delta-V overhaul, and all this work is drifting and merge conflicting as
I rename things, so I'm going to merge the suicide burn timer a bit
half-done.  The module isn't wired up anywhere so shouldn't run.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>