Change some RCS "display" logic
If you have a stage which has RCS and a thruster that both pull the
same fuel (e.g. MMH+NTO) then the "RCS ∆v max value" which is really
the value-after-all-the-normal-engine-fuel-tanks-are-drained will be
zero because the thruster drained all the shared RCS fuel.

So use the "min" values for stuff like resources burned and deltatime
and the showEmpty button should key off the min value instead of max.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>