Fix DeltaV Stats and planes in RF
When RF is loaded ModuleEngines can still be non-RF module engines
so all the reflection needs to be conditional on checking we've
got a valid ModuleEnginesRF.

Old code used to do this by handling exceptions, new code sets up a
boolean to check the type dynamically.  On build we take the hit
of doing expensive dynamic type reflection, but tick to tick in
update we use the boolean.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>