Better PVG+StageStats integration
Save off the initial conditions at the time that the stage stats
simulation starts running.  Then when the results of that are fed into
PVG, use those initial conditions so that the PVG simulation is
consistent and not using stagestats results from one tick and initial
conditions from another.

This should reduce visual 'flicker' in the predictions display at the
very least.  It might increase accuracy.

It does allow the update cycle of stage stats to be tweaked for
performance without really affecting PVG accuracy.  PVG producing
results that are a few 100ms stale doesn't really matter since it
produces the whole interpolated future trajectory.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>