Tweak DecouplingAnalyzer algorithm
As we walk through the vessel we should never decrease the
decoupledInStage value to a lower value based on the inverseStage of
the decoupler.

If that happens then what we've done is fire a decoupler that is already
decoupled.  In that case the decoupler is rendered "sterile" and is
treated like just another part.

This is a departure from old behavior, but is necessary to get some
analysis to be correct, and to agree with the autostaging behavior of
mechjeb (and to make PVG work correctly).