Hohmann/Bi-impulsive overhaul
The problem has been recrafted as a conjugate gradient problem with
a bit of heuristics for a search.  Limiting to coplanar is still
supported.  Transfer/Rendezvous is still supported.

There are buttons for no capture burn and to plan the capture burn
(maybe that should be "intercept burn" or "matching burn" or
something?), but they don't yet work.

The period offset thingy needs to be changed and doesn't yet work.

The simulated annealing is gone for now and will probably come back as
a rewriting with a basin hopping algorithm, but needs some UI work.

This is going to be best when dealing with things that are closer to
circular-to-circular coplanar transfer with smaller synodic periods, and
will still break as you get away from it.

It can at least plan transfer burns to GEO with the optimal inclination
change at the equator included (~2.16 degrees for a 185x185 launch
from Kennedy).  All that is lacking is a UX/UI way to construct an
artificial target.orbit.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>