Behavioral changes to Node Executor
The stock node executor no longer "hunts" for the last dribbling
bit of the maneuver node and the "tolerance" setting has been
eliminated.  There is a 1 second terminal guidance period where
tracking is frozen and the burn terminates when the angle between
the heading and the maneuver node is >= 90 degrees.

This also backs out the MANEUVER_COT and thrustForward stuff for
asymmetric thrust vehicles.  This is because thrustForward isn't
valid if the engines are off.  Don't know if VesselState needs
fixing to fix thrustForward or if the node executor needs to
switch modes based on thrust being on or not, but it feels like
VesselState should be updated.

replaces #1771

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>