Simplify RCS ullage
Always apply RCS ullage if we're less than verystable.

Don't kill the engine if the last command was to apply throttle.

In my head, this means that if we drop below verystable during spoolup we
should apply some RCS, without having logic that will accidentally kill
the engine and waste an ignition if ullage drops before spoolup is done.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>