Add back a minimum rcs time and ignition delay
This is working around the fact that "very stable" still has a
0.2% chance of failure and that we want to deliberately "over ullage"
the engine, and we want to wait for a couple of ticks above
"very stable" to ensure we are at 100% before engaging the throttle.

Ullage may still bounce down below "very stable" again, but we'll
reapply ullage as soon as it does.

RF does not currently expose any API that allows seeing if we're at
an ullage state over 100%, we might be able to directly look at
propellantStability to see if we're actually at 100% or only at 99.6%

It would be nice if RF would not truncate propellantStability at 100%
and would have Stable be 100% and VeryStable be something like 105%.

(And note that currently the exponent of RF is 0.5 so that 99.6% ullage
results in a 99.8% chance of ignition)

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>