Overhaul of ReturnFromMoon initial guesser
It now uses Brent's minimization algorithm to try to guess the
inclination of the return ellipse and if it is type I or type II
return orbit.

Restored the 1/2 time midpoint constraints since that helps
convergence speed in the hyperbolic initial conditions case and
seems to improve other ones as well.

This also now finds some better solutions as evidenced by the
one test which dropped from 1520 m/s to 1092 m/s.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>