PID rewrite
Adds PIDLoop2 which is written more conventionally.

- exposes pterm, iterm and dterm for inspection
- trapezoidal integration/differentiation
- standard and parallel form setters
- separate out the deadbands for p, i, d and output
- proper tracking anti-windup
- first order reset element behavior (extended clegg integrator)

Also fixes bugs in the B and C options for 2DOF behavior.

Tweaks the BetterController to just be a cascade with a PI instead of
the PID, with a more aggressive integral term and no derivative.  Backed
off the p-term over what matlab recommended, because it made some
engines with too much gimbal way too twitchy still.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>