Fix PVG settings stage numbers
When multiple stages in the Δv stats window have the same KSP stage
number, the (unguided) and (optimize) labels get applied to the wrong
stage(s). This commit changes the window to apply the labels to the
stage with the correct KSP stage, which may not be the same as the index
in the vacStats list.

Fix #1920.