Fix coasts after the early shutoff stage
When a coast is inserted immediately after the "Early shutoff stage",
RCS does not get enabled and the `StartCoast` variable doesn't get set;
this means that the coast time does not tick down.

This patch makes it so both those things do get done when starting the
coast after the early shutoff stage.