Fix a problem where coast time doesn't count down
When 'Coast after' is checked, the kspStage inside the loop is one past
the vessel's current stage. Similarly, if autostaging hasn't staged to
the point where the vessel's current stage is the coast stage, then this
check will also fail. In either case the coast time will not tick
down.

If we are coasting, then the timers should always be ticking down,
whatever the current stage happens to be. Therefore, the
Vessel.currentStage check is unnecessary.