Fix Suicide Burntime calculation
This is an analytic function (inside a root solver) which assumes
constant twr and flat constant surface gravity.

It will not be perfect, but it should be 'safe' in that if you
start your burn before zero you should not collide with the surface.

It tries to accurately predict the height of the terrain at the predicted
downrange burnout point, but extremely rugged terrain (e.g. cliffs)
may void this warranty.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>