Fix classic launch inclination
I don't know what I thought I was doing in the Launch inclination stuff,
but this mostly fixes it.  It drives the inclination error to zero as
the vgo to raise the apoapsis goes to zero via proportional steering.

There still needs to be some terminal guidance handling, right now it
gets all weird and spinny as the dv goes to zero.

Needs to have the AoA limiter turned off or the fadeout very high for
accuracy.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>