Update BetterController to have a real PID for the position
This will allow adding an integral term to the pos PID, and
possibly a 2DOF PID there.

This shouldn't change much behavior, all the PID settings are
the same.

I snuck in a TorqueSmooth tunable which is a low pass filter over
the torque from VesselState which can help with wiggles from
control surfaces near the edge of the atmosphere (KSP and/or
FAR is doing something weird).

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>