Fix MissingMethodException in ascent guidance
While `Vector3d` has the `RotateTowards` method declared in its API,
there's no implementation for it; it's declared as a "runtime builtin"
method, but the runtime does not provide this method (unlike the
`Vector3` version, which does have an implementation in the Unity
runtime).

The result is a `MissingMethodException` whenever this method is called.

For reference, the same issue exists with the `Slerp` and
`Orthonormalize` methods, which also will not work.