Add DirectionTracker and wire it up to BetterController
The DirectionTracker is my attempt to solve the problem of splitting
up desired vs current pitch/yaw/roll so that I can implement LQRs and
2DOF PID controllers.  It attempts to track small tick-to-tick
deltas and use them to accumulate a value which will not suffer from
gimbal lock issues or 2pi wraparound issues.  When the controller axis
is reset then in addition to clobbering the PID integral states, it also
zeros the DirectionTracker.  Since the error calc is still mathematically
identical at the end of the day it really shouldn't change the
BetterController much (i hope).

This also adds an LQR based controller which is just a proof of concept
that I can actually wire up an LQR controller.  It also uses the
DirectionTracker.  It does not usably fly a rocket in its current form.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>