Run Qα limiter during pitch program again
This reverts to previous behavior.

I tried fixing it so that zero Qα would work, but that would
take more detailed patch to bypass Qα during initiation, but
not pitchover, being careful that it would be smooth and not
ratchet (so it would initiate by 3 degrees or so, but
then not apply a zero Qα limit and try to go back to nearly vertical,
etc).

Didn't have an actual bug to fix that behavior, so the bugs the
fix introduced and got reported were worse.