Add Cd and Aref boxes to PSG settings
Aref is set to zero initially so aero will be turned off unless
you "turn it on" by putting the correct value in for the rocket.

I believe the only way to find the correct values for these is
going to be FAR analysis?

But figuring it out by hand based on the diameter of the rocket
and assuming 0.5 for Cd should work reasonably well.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>